Reduce length of filenames stored in rlibs
Previously we would store
    <filestem>.<crate_name>.<stable_crate_id>-cgu.<N>.rcgu.o
where <filestem> is generally the crate name + extra-filename. With this
change we only store the less redundant
    <crate_name>.<stable_crate_id>-cgu.<N>.rcgu.o
This still keeps the codegen unit name unique within a dependency graph
as the stable crate id is enforced to be unique within this scope. We
can't skip the filestem part for the temporary object files as that
could cause two independent compilations of two crates with the same
name and stable crate id to clobber each other's temporary files even
when they never get linked together.